# docs/swagger.yaml
definitions:
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /manga/{id}/rating-stats:
    get:
      description: Return average rating, total review count, and distribution for
        a manga
      parameters:
      - description: Manga ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Rating stats
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get rating statistics
      tags:
      - reviews
  /manga/{id}/reviews:
    get:
      description: List all reviews for a manga with pagination
      parameters:
      - description: Manga ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Results per page (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reviews list
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get manga reviews
      tags:
      - reviews
    post:
      consumes:
      - application/json
      description: Submit a rating and review text for a manga
      parameters:
      - description: Manga ID
        in: path
        name: id
        required: true
        type: string
      - description: 'Review payload — {rating: int (1-10), text: string}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Review created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Invalid rating or text
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Create a review
      tags:
      - reviews
...
  /reviews/{review_id}:
    delete:
      description: Permanently remove your own review
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Review deleted
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Not your review
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a review
      tags:
      - reviews
    get:
      description: Retrieve a single review by its ID
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Review data
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Review not found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get a review
      tags:
      - reviews
    put:
      consumes:
      - application/json
      description: Edit the rating or text of your own review
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: string
      - description: 'Updated review — {rating: int (1-10), text: string}'
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Review updated
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Invalid rating
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "403":
          description: Not your review
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a review
      tags:
      - reviews
  /reviews/{review_id}/helpful:
    post:
      description: Increment the helpful vote counter on a review
      parameters:
      - description: Review ID
        in: path
        name: review_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Helpful count updated
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Mark review as helpful
      tags:
      - reviews
...
  /users/reviews:
    get:
      description: List all reviews written by the authenticated user
      parameters:
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Results per page (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User's reviews
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Get my reviews
      tags:
      - reviews
swagger: "2.0"